Move tango object server to its own module
